change user available points on add and remove order. Change points type from BigDecimal to long

+++ com.sap.hana.cloud.samples.benefits/src/com/sap/hana/cloud/samples/benefits/api/bean/BenefitItemBean.java
@@ -1,7 +1,5 @@
 package com.sap.hana.cloud.samples.benefits.api.bean;
 
-import java.math.BigDecimal;
-
 import com.google.gson.annotations.Expose;
 import com.sap.hana.cloud.samples.benefits.persistence.model.OrderDetails;
 
@@ -17,7 +15,7 @@ public class BenefitItemBean {
 	public long quantity;
 	
 	@Expose
-	public BigDecimal itemValue;
+	public long itemValue;
 
 	public void init(OrderDetails orderItem) {
 		this.id = orderItem.getId();
